+++ src/hfsinspect.c
@@ -666,7 +666,7 @@ OPEN:
         unsigned     sectorCount = 16;
         size_t       readSize    = vol->sector_size*sectorCount;
         void*        buf         = NULL;
-        int          nbytes      = 0;
+        ssize_t      nbytes      = 0;
 
         SALLOC(buf, readSize);
         assert(buf != NULL);
